+++ src/view/__init__.py
@@ -1,5 +1,4 @@
-from .main_view import MainWindow
-from .main_widget import MainWidget
+from .main_view import (MainWindow, MainWidget)
 from .file_syncronized_widget import FileSyncronizedWidget
 from .lateral_menu_widget import LateralMenuWidget
 from .settings_widget import SettingsWidget
